requirements.txt
transformers==4.32.0
accelerate
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
torchvision
pillow
tensorboard
matplotlib
git+https://github.com/huggingface/peft
deepspeed
